requirements.txt
requirements.txt
functions-framework
requests
pandas
dlt
google-cloud-secret-manager
google-cloud-bigquery
